+++ ui/e2e/qa-uday-17may2026.spec.ts
@@ -1,4 +1,4 @@
-import { expect, test } from "@playwright/test";
+import { expect, test } from "./helpers/test";
 import { authenticate } from "./helpers/auth";
 
 /**